% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{cardinalize}
\alias{cardinalize}
\title{internal fn to create cardinal rank scores}
\usage{
cardinalize(x)
}
\arguments{
\item{x}{input vector}
}
\value{
returns new vector of cardinal rank scores
}
\description{
internal function for generating cardinal ranks
}
\details{
converts raw Elo scores into predicted number of individuals beaten
  (using Equation 1 from paper)

subtracting .5 is equivalent to removing the prob of winning against oneself
  because 1/(1 + exp(-0.01*0)) = 1/(1 + exp(0)) = 1/(1 + 1) = 1/2
}
