% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chimpagg_f}
\alias{chimpagg_f}
\title{Anonymized female chimpanzee pant-grunt data from Gombe}
\format{
A data frame with 1015 rows and 3 variables:
\describe{
  \item{Date}{date of interaction}
  \item{Winner}{winning individual}
  \item{Loser}{losing individual}
}
}
\source{
Supplemental data published with Foerster, Franz et al. (2016).
  \url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.r4g74}
}
\usage{
chimpagg_f
}
\description{
Female data from Gombe National Park, Tanzania from 1969 to 2013.  
  Data are submissive pant-grunt vocalizations.
}
\keyword{datasets}
