% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chimpagg_m}
\alias{chimpagg_m}
\title{Anonymized male chimpanzee pant-grunt data from Gombe}
\format{A data frame with 2741 rows and 3 variables:
\describe{
  \item{Date}{date of interaction}
  \item{Winner}{winning individual}
  \item{Loser}{losing individual}
}}
\source{
Supplemental data published with Foerster, Franz et al. (2016).
  \url{https://datadryad.org/resource/doi:10.5061/dryad.r4g74}
}
\usage{
chimpagg_m
}
\description{
Data from Gombe National Park, Tanzania from 1978 to 2011.  Data 
  are submissive pant-grunt vocalizations.
}
\keyword{datasets}
