% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chimppres_f}
\alias{chimppres_f}
\title{Anonymized female chimpanzee presence data from Gombe}
\format{
A data frame with 44 rows and 3 variables:
\describe{
  \item{id}{female code}
  \item{start_date}{start date}
  \item{end_date}{date of departure}
}
}
\source{
Supplemental data published with Foerster, Franz et al. (2016).
  \url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.r4g74}
}
\usage{
chimppres_f
}
\description{
Female presence data from Gombe National Park, Tanzania from 1969 to 2013.  
  Presence criteria are given in Foerster, Franz et al. (2016)
}
\keyword{datasets}
