% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chimppres_m}
\alias{chimppres_m}
\title{Anonymized male chimpanzee presence data from Gombe}
\format{
A data frame with 22 rows and 3 variables:
\describe{
  \item{id}{male code}
  \item{start_date}{start date}
  \item{end_date}{date of departure}
}
}
\source{
Supplemental data published with Foerster, Franz et al. (2016).
  \url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.r4g74}
}
\usage{
chimppres_m
}
\description{
Male presence data from Gombe National Park, Tanzania from 1978 to 2011.  
  Presence criteria are given in Foerster, Franz et al. (2016)
}
\keyword{datasets}
