% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.m3_lik_vect.R
\name{elo.m3_lik_vect}
\alias{elo.m3_lik_vect}
\title{optimize k parameter and entry Elo scores, vectorized}
\usage{
elo.m3_lik_vect(par, IA_data, all_ids)
}
\arguments{
\item{par}{list of parameters, with par[1] being log(k), and par[2:length(par)] 
being the initial elo scores of individuals}

\item{IA_data}{list of interaction data, with columns "Date", "Winner", and "Loser"
(in that order)}

\item{all_ids}{list of all ids to rank}
}
\description{
Function to optimize k parameter and entry Elo scores
}
\examples{

# for internal use
}
