% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.model3.R
\name{elo.model3}
\alias{elo.model3}
\title{Function to optimize k parameter and entry Elo scores}
\usage{
elo.model3(par, IA_data, all_ids, return_likelihood = T)
}
\description{
Function to optimize k parameter and entry Elo scores
}

