% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EloOptimized.R
\docType{package}
\name{EloOptimized}
\alias{EloOptimized}
\alias{EloOptimized-package}
\title{EloOptimized: ML fitting of Elo Scores}
\description{
This package implements the maximum likelihood methods for deriving Elo scores as published in Foerster, Franz et al. (2016).
Chimpanzee females queue but males compete for social status. Scientific Reports 6, 35404, doi:10.1038/srep35404
}
\section{Primary functions}{

\itemize{
  \item{\code{\link{eloratingopt_simple}}: main function}
  \item{\code{\link{elo.model1}}: internal function for fitting model type 1}
  \item{\code{\link{elo.model3}}: internal function for fitting model type 3}
  \item{\code{\link{elo.m3_lik_vect}}: vectorized internal function 
    for fitting mod type 3}
}
}

\section{Ideas for future development}{

\itemize{
  \item option to include additional rank scores or not (ie leave out ordinal, scaled, etc...)
  \item rather than returning a data frame, we could return a fit object with $daily_elo, 
    $optimized_k, $prediction_accuracy, $log_likelihood, $aic, $aicc, etc...
  \item maybe name the models by model type, rather than M or F?  That way we're not presupposing
    that males all have the same entry value, etc...
  \item option to go back and recalculate during burn-in period with optimized k for Model 1?
  \item find data we can use in vignette
  \item create vignette, other package doohickies
}
}

