% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloratingopt.R
\name{eloratingopt}
\alias{eloratingopt}
\title{Get optimal Elo ratings and save rank file}
\usage{
eloratingopt(agofile, presencefile, sex, outputfile)
}
\arguments{
\item{agofile}{Input file with dominant interactions.}

\item{presencefile}{Input file with date in rows, individuals in columns, and 0/1 for absent/present}

\item{sex}{Whether data are for males "M" or females "F", no default}

\item{outputfile}{Name of file to save ranks to}
}
\description{
Conducts optimized elo rating analyses as per Foerster, Franz et al
and saves raw, normalized, cardinal, and  categorical ranks in output file.
}
\examples{
eloratingopt(agofile="males_ago.csv", presencefile="male_presence.csv", sex="M") 
}

