% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloratingtrad.R
\name{eloratingtrad}
\alias{eloratingtrad}
\title{Create daily traditional elo ranks and multiple derivatives}
\usage{
eloratingtrad(agon_data, pres_data, k = 100, init_elo = 1000, outputfile = NULL, 
  returnR = TRUE)
}
\arguments{
\item{agon_data}{Input data frame with dominance interactions, should only contain Date, 
Winner, Loser.  Date should be formatted as MONTH/DAY/YEAR, or already as Date class.}

\item{pres_data}{Input data frame with columns "id", "start_date" and "end_date".  Date
columns should be formatted as MONTH/DAY/YEAR, or already as Date class.  If all IDs 
are present the whole time, you ignore this and a pres_data table will be automatically
generated.}

\item{k}{Specified value of the k parameter, default is 100}

\item{init_elo}{The starting Elo value for all individuals, default is 1000}

\item{outputfile}{Name of csv file to save ranks to.  Default is NULL, in which case 
the function will only return a table in R.  If you supply an output file name
the function will save the results as a csv file in your working directory.}

\item{returnR}{whether to return an R object from the function call.  Default is TRUE}
}
\description{
Conducts traditional elo rating analyses using specified K value
  and saves raw, normalized, cardinal, and  categorical ranks in output file.
}
\examples{

nbadata = EloOptimized::nba #nba wins and losses from the 1995-96 season
nbaelo = eloratingtrad(agon_data = nbadata)
# generates traditional Elo scores (with init_elo = 1000 & k = 100) and saves 
#   them as "nbaelo" 

}
