% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{jenksify}
\alias{jenksify}
\title{internal fn to generate categorical ranks}
\usage{
jenksify(x)
}
\arguments{
\item{x}{input vector}
}
\value{
returns new vector of categorical ranks (high/medium/low)
}
\description{
internal function for generating categorical ranks using 
  jenks natural breaks algorithm
}
\details{
creates categorical ranks using jenks natural breaks algorithm
}
