% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nba}
\alias{nba}
\title{NBA games 1995-96}
\format{
A data frame with 1189 rows and 3 variables:
\describe{
  \item{Date}{date of game}
  \item{Winner}{winning team}
  \item{Loser}{losing team}
}
}
\source{
\url{https://github.com/fivethirtyeight/data/blob/master/nba-elo/nbaallelo.csv}
}
\usage{
nba
}
\description{
Outcome of NBA games during the 1995-1996 regular season, adapted from a dataset 
  from fivethirtyeight
}
\keyword{datasets}
