% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_fns.R
\name{relativize}
\alias{relativize}
\title{internal fn to relativize rank scores}
\usage{
relativize(x)
}
\arguments{
\item{x}{input vector}
}
\value{
returns new vector of scaled rank scores
}
\description{
internal function for generating scaled cardinal ranks
}
\details{
scales cardinal Elo scores between 0 and 1
}
