% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo.model1.R
\name{elo.model1}
\alias{elo.model1}
\title{Function to optimize k parameter in Elo Rating Method}
\usage{
elo.model1(par, burn_in = 100, init_elo = 1000, IA_data, all_ids,
  return_likelihood = T)
}
\description{
Function to optimize k parameter in Elo Rating Method
}
