% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eloratingopt.R
\name{eloratingopt}
\alias{eloratingopt}
\title{Create daily elo ranks and multiple derivatives (simplified)}
\usage{
eloratingopt(agon_data, pres_data, mod_type, outputfile = NULL, returnR = TRUE)
}
\arguments{
\item{agon_data}{Input data frame with dominance interactions, should only contain Date, 
Winner, Loser.  Date should be formatted as MONTH/DAY/YEAR, or already as Date class.}

\item{pres_data}{Input data frame with columns "id", "start_date" and "end_date".  Date
columns should be formatted as MONTH/DAY/YEAR, or already as Date class.  If all IDs 
are present the whole time, you ignore this and a pres_data table will be automatically
generated.}

\item{mod_type}{1 to fit only K, 3 to fit K and starting Elo for each individual.
MUCH slower for mod_type 3.}

\item{outputfile}{Name of csv file to save ranks to.  Default is NULL, in which case 
the function will only return a table in R.  If you supply an output file name
the function will save the results as a csv file in your working directory.}

\item{returnR}{whether to return an R object from the function call.  Default is TRUE}
}
\description{
Conducts optimized elo rating analyses as per Foerster, Franz et al
  and saves raw, normalized, cardinal, and  categorical ranks in output file.
}
\examples{

nbadata = EloRatingOptimized::nba #nba wins and losses from the 1995-96 season
nbaelo = eloratingopt(agon_data = nbadata, mod_type = 1)
# generates optimized elo scores (optimizing only K) and saves them as "nbaelo" 

}
